ocr/requirements.txt
gradio
google-generativeai
pillow
python-dotenv